pyproject.toml
[tool.poetry]
name = "image-validator-app"
version = "2.0.0"
description = ""
authors = ["Vety <vetylestari3@gmail.com>"]
readme = "README.md"
packages = [
    { include = "project" },
]

[tool.poetry.dependencies]
python = ">=3.9, <3.13"
fastapi = "^0.112.0"
uvicorn = {extras = ["standard"], version = "^0.30.5"}
python-multipart = "^0.0.9"
pillow = "^10.4.0"
torch = "^2.3.1"
torchvision = "^0.19.0"
psutil = "^6.1.1"

[poetry.group.dev.dependencies]
pytest = "^6.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
fastapi = "handler_fastapi:run"